backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.129.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
numpy==2.4.2
openai==2.21.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
